livekit-agent/requirements.txt
# LiveKit Agents
livekit-agents>=0.8.0
livekit-agents-openai>=0.8.0
livekit-agents-elevenlabs>=0.8.0
livekit-agents-assemblyai>=0.8.0

# Core dependencies
asyncio
aiohttp>=3.8.0
websockets>=11.0

# AI/ML
openai>=1.0.0
elevenlabs>=0.2.0

# Database
supabase>=2.0.0
qdrant-client>=1.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0

# Logging
structlog>=23.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
